controller update

+++ Rest/src/main/java/app/cos/rest/dto/response/CowAlertDTO_response.java
@@ -3,7 +3,7 @@ package app.cos.rest.dto.response;
 public class CowAlertDTO_response {
 
 	private Long id;
-	private int id_cow;	
+	private Long id_cow;	
 	private int bcs_threshold_min;
 	private int bcs_threshold_max;
 	
@@ -13,10 +13,10 @@ public class CowAlertDTO_response {
 	public void setId(Long id) {
 		this.id = id;
 	}
-	public int getId_cow() {
+	public Long getId_cow() {
 		return id_cow;
 	}
-	public void setId_cow(int id_cow) {
+	public void setId_cow(Long id_cow) {
 		this.id_cow = id_cow;
 	}
 	public int getBcs_threshold_min() {
